factor unmarshal code into unmarshal.go
